# Performance optimizations for SvgNumberCollectionConverter parser