--- a/Tests/Svg.Benchmark/SvgNumberCollectionConverterBenchmarks.cs
+++ b/Tests/Svg.Benchmark/SvgNumberCollectionConverterBenchmarks.cs
@@ -1,0 +1,15 @@
+using System;
+using BenchmarkDotNet.Attributes;
+
+namespace Svg.Benchmark
+{
+    public class SvgNumberCollectionConverterBenchmarks
+    {
+        [Benchmark]
+        public void SvgNumberCollectionConverter_Parse()
+        {
+            SvgNumberCollectionConverter.Parse("1.6 3.2 1.2 5".AsSpan());
+            SvgNumberCollectionConverter.Parse("1.6,3.2,1.2,5".AsSpan());
+        }
+    }
+}